+++ src/scripts/components/stories/header/header.tsx
@@ -2,9 +2,6 @@ import React, {FunctionComponent} from 'react';
 
 import Button from '../../main/button/button';
 import {ArrowBackIcon} from '../../main/icons/arrow-back-icon';
-import {useStoryParams} from '../../../hooks/use-story-params';
-
-import {SlideType} from '../../../types/slide-type';
 
 import styles from './header.styl';
 
@@ -19,25 +16,18 @@ const Header: FunctionComponent<Props> = ({
   title,
   backButtonId,
   children
-}) => {
-  const {selectedStory, slideIndex} = useStoryParams();
-
-  const isSplashScreen =
-    selectedStory?.slides[slideIndex].type === SlideType.Splashscreen;
-
-  return (
-    <div className={styles.header}>
-      <Button
-        className={styles.backButton}
-        icon={ArrowBackIcon}
-        label={backButtonId}
-        link={backLink}
-        hideLabelOnMobile
-      />
-      {!isSplashScreen && <h1 className={styles.title}>{title}</h1>}
-      <div className={styles.rightContent}>{children}</div>
-    </div>
-  );
-};
+}) => (
+  <div className={styles.header}>
+    <Button
+      className={styles.backButton}
+      icon={ArrowBackIcon}
+      label={backButtonId}
+      link={backLink}
+      hideLabelOnMobile
+    />
+    {title && <h1 className={styles.title}>{title}</h1>}
+    <div className={styles.rightContent}>{children}</div>
+  </div>
+);
 
 export default Header;