+++ src/scripts/config/main.ts
@@ -80,7 +80,14 @@ export default {
     twitter:
       'http://twitter.com/intent/tweet?text=ESA%20Climate%20From%20Space&url={currentUrl}'
   },
+  planeratyVisionsLogo: 'assets/images/planetary-visions.png',
+  ubilabsWebsite: 'https://ubilabs.net',
+  planetaryVisionsWebsite: 'http://planetaryvisions.com/',
+  githubRepo: 'https://github.com/ubilabs/esa-climate-from-space',
+  cciWebsite: 'https://climate.esa.int/',
+  esaWebsite: 'https://www.esa.int/',
   legendImage: `${baseUrlTiles}/{id}/legend.png`,
   downloadUrls,
-  localStorageLanguageKey: 'language'
+  localStorageLanguageKey: 'language',
+  delay: 5000
 };